refactor(utils): extract complexity scoring to dedicated utility
