+++ src/core/reporters/MarkdownReporter.ts
@@ -40,6 +40,7 @@ import {
   groupWebResourcesByType,
   groupCustomAPIsByBinding,
 } from '../utils/grouping.js';
+import { calculateComplexityScore } from '../utils/complexity.js';
 
 export class MarkdownReporter {
   /**
@@ -2247,7 +2248,7 @@ export class MarkdownReporter {
 
     // Calculate complexity for each entity
     const scores = result.entities.map(entity => {
-      const score = this.calculateComplexityScore(entity, result);
+      const score = calculateComplexityScore(entity, result);
       return {
         entity: entity.entity.LogicalName,
         displayName: entity.entity.DisplayName?.UserLocalizedLabel?.Label || entity.entity.LogicalName,
@@ -2638,46 +2639,6 @@ export class MarkdownReporter {
   }
 
 
-  /**
-   * Calculate complexity score for entity
-   */
-  private calculateComplexityScore(entity: EntityBlueprint, _result: BlueprintResult): {
-    total: number;
-    level: 'Low' | 'Medium' | 'High';
-    breakdown: {
-      attributes: number;
-      plugins: number;
-      flows: number;
-      businessRules: number;
-      forms: number;
-    };
-  } {
-    const breakdown = {
-      attributes: entity.entity.Attributes?.length || 0,
-      plugins: entity.plugins.length,
-      flows: entity.flows.length,
-      businessRules: entity.businessRules.length,
-      forms: entity.forms.length,
-    };
-
-    const total =
-      breakdown.attributes * 1 +
-      breakdown.plugins * 5 +
-      breakdown.flows * 3 +
-      breakdown.businessRules * 2 +
-      breakdown.forms * 2;
-
-    let level: 'Low' | 'Medium' | 'High';
-    if (total <= 50) {
-      level = 'Low';
-    } else if (total <= 150) {
-      level = 'Medium';
-    } else {
-      level = 'High';
-    }
-
-    return { total, level, breakdown };
-  }
 
   /**
    * Generate security overview with special permissions matrix
